Embedded: Moved authoring html and controller functions to separate authoring files. Closes #1290
